Fix web tests
