+++ packages/flutter/test/native/cocoa_envelope_sender_test_web.dart
@@ -1,0 +1,10 @@
+// Stub for web - these tests only run on VM
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('Cocoa envelope sender tests are not supported on web', () {
+    // This test file exists only to satisfy the compiler when running web tests.
+    // The actual tests in cocoa_envelope_sender_test_real.dart are only
+    // executed on VM platforms.
+  });
+}
